research-digest/requirements.txt
flask==3.1.0
flask-login==0.6.3
flask-sqlalchemy==3.1.1
werkzeug==3.1.3
langchain==0.3.14
langchain-community==0.3.14
langchain-ollama==0.2.3
faiss-cpu
sentence-transformers==3.3.1
feedparser==6.0.11
scikit-learn==1.6.1
apscheduler==3.11.0
loguru==0.7.3
pyyaml==6.0.2
numpy>=2.1.0
beautifulsoup4==4.12.3
lxml==5.3.0
requests==2.32.3
python-dotenv==1.0.1
